requirements.txt
fastapi==0.108.0
uvicorn==0.25.0
pydantic==2.5.2
groq==0.4.0
python-dotenv==1.0.0
requests==2.31.0
mcp==0.1.0
discord.py==2.3.2
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
numpy==2.2.4
pandas==2.2.3
pydantic-settings==2.8.1
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
sniffio==1.3.1
sse-starlette==2.2.1
starlette==0.46.1
typing_extensions==4.13.0
tzdata==2025.2
urllib3==2.3.0 